requirements.txt
fastapi
uvicorn
aiohttp
Pillow
supabase
python-multipart
